[refactor] MVC 구조에 맞게 패키지 이동
