+++ src/main/java/project/game/domain/MoveRule.java
@@ -1,4 +1,4 @@
-package project.game;
+package project.game.domain;
 
 public interface MoveRule {
     boolean canMove();
