+++ src/main/java/project/game/domain/RandomNumberProvider.java
@@ -1,4 +1,4 @@
-package project.game;
+package project.game.domain;
 
 public interface RandomNumberProvider {
     int getRandomNumber();
