+++ src/main/java/project/game/view/ConsoleView.java
@@ -1,4 +1,4 @@
-package project.game;
+package project.game.view;
 
 import project.InputTool;
