+++ src/main/java/project/game/view/View.java
@@ -1,4 +1,4 @@
-package project.game;
+package project.game.view;
 
 import java.util.List;
