+++ src/test/java/project/game/RandomTest.java
@@ -2,6 +2,8 @@ package project.game;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.RepeatedTest;
+import project.game.domain.RandomNumberProvider;
+import project.game.domain.RandomNumberProviderImpl;
 
 import static org.assertj.core.api.Assertions.assertThat;
